switched from AppleScript to JXA
AppleScript loops very slowly: processing a character in the input takes
longer as the input grows! Seems to be O(n^2) or something for a simple
loop.

JXA is somewhat less well documented, but it's Javascript, and loops fast.

All of the same fixes are needed, and I discovered a couple more:

+ and * inside of vim, inside a PuTTY/RDP session (aka: beyondtrust).
	vim tries to interpret them somehow, so we send raw keycodes now.